Rename App to ScoreApp
Minimum file renaming and code changes needed to keep the app working
